Step 10.4: Update cache on message added
