+++ src/components/ChatRoomScreen/index.tsx
@@ -1,4 +1,3 @@
-import { defaultDataIdFromObject } from 'apollo-cache-inmemory';
 import gql from 'graphql-tag';
 import React from 'react';
 import { useCallback } from 'react';
@@ -8,8 +7,8 @@ import MessageInput from './MessageInput';
 import MessagesList from './MessagesList';
 import { History } from 'history';
 import { useGetChatQuery, useAddMessageMutation } from '../../graphql/types';
-import * as queries from '../../graphql/queries';
 import * as fragments from '../../graphql/fragments';
+import { writeMessage } from '../../services/cache.service';
 
 const Container = styled.div`
   background: url(/assets/chat-background.jpg);
@@ -41,10 +40,6 @@ const addMessageMutation = gql`
 interface ChatRoomScreenParams {
   chatId: string;
   history: History;
-}
-
-interface ChatsResult {
-  chats: any[];
 }
 
 const ChatRoomScreen: React.FC<ChatRoomScreenParams> = ({
@@ -79,75 +74,7 @@ const ChatRoomScreen: React.FC<ChatRoomScreenParams> = ({
           },
         },
         update: (client, { data: { addMessage } }) => {
-          type FullChat = { [key: string]: any };
-          let fullChat;
-          const chatIdFromStore = defaultDataIdFromObject(chat);
-
-          if (chatIdFromStore === null) {
-            return;
-          }
-
-          try {
-            fullChat = client.readFragment<FullChat>({
-              id: chatIdFromStore,
-              fragment: fragments.fullChat,
-              fragmentName: 'FullChat',
-              data: fullChat,
-            });
-
-            let data;
-            try {
-              data = client.readQuery<ChatsResult>({
-                query: queries.chats,
-              });
-            } catch (e) {
-              return;
-            }
-
-            if (!data || data === null) {
-              return null;
-            }
-            if (!data.chats || data.chats === undefined) {
-              return null;
-            }
-            const chats = data.chats;
-
-            const chatIndex = chats.findIndex((c: any) => c.id === chatId);
-            if (chatIndex === -1) return;
-            const chatWhereAdded = chats[chatIndex];
-
-            // The chat will appear at the top of the ChatsList component
-            chats.splice(chatIndex, 1);
-            chats.unshift(chatWhereAdded);
-
-            client.writeQuery({
-              query: queries.chats,
-              data: { chats: chats },
-            });
-          } catch (e) {
-            return;
-          }
-
-          if (!data || data === null) {
-            return null;
-          }
-          if (!data.chats || data.chats === undefined) {
-            return null;
-          }
-          const chats = data.chats;
-
-          const chatIndex = chats.findIndex((c: any) => c.id === chatId);
-          if (chatIndex === -1) return;
-          const chatWhereAdded = chats[chatIndex];
-
-          // The chat will appear at the top of the ChatsList component
-          chats.splice(chatIndex, 1);
-          chats.unshift(chatWhereAdded);
-
-          client.writeQuery({
-            query: queries.chats,
-            data: { chats: chats },
-          });
+          writeMessage(client, addMessage);
         },
       });
     },
